requirements.txt
snowflake-snowpark-python>=1.5.0
snowflake-connector-python>=3.0.0
pandas>=1.3.0 